Create GLJob system to send commands to main thread
In this commit, I have devised a method of queueing OpenGL commands from
worker threads to the main thread with OpenGL context, instead of
passing the OpenGL context to the threads that requires it. This
prevents unnecessary race condition problems and increases performance
of the main rendering thread.

OpenGL commands are passed as a lambda function to a queue in the main
thread which executes 1 function per frame.