Model can finally create and save image embeddings
Had a bug in CLIP implementation where newly allocated memory would have
some NaN values and cause errors furthur in the model.

With ggml, if a tensor is allocated in the compute graph, either it will
be assigned an input value, a weight/bias or some preset value.

Problem: The `embedding` tensor in compute graph was not assigned any
value and was picking up some NaN values, causing errors.
Solution: Set it to 0 (It was in the original implementation also).